Fix E2E test file import and package.json path #30
