+++ src/routes/page.e2e.ts
@@ -1,39 +1,12 @@
 import { readFileSync } from 'node:fs'
-import { expect, test, type Page } from '@playwright/test'
+import { expect, test } from '@playwright/test'
 
 const { version } = JSON.parse(
-	readFileSync(new URL('../package.json', import.meta.url), 'utf-8'),
+	readFileSync(new URL('../../package.json', import.meta.url), 'utf-8'),
 ) as { version: string }
 
 const LOADING_OVERLAY_TIMEOUT_MS = 8000
 const READY_PROGRESS_VALUE = 100
-const TOUCH_PRIMARY_QUERY = '(hover: none) and (pointer: coarse)'
-
-export async function stub_touch_primary(page: Page, is_touch: boolean): Promise<void> {
-	await page.addInitScript(
-		([query, matches]) => {
-			const original = globalThis.matchMedia.bind(globalThis)
-			globalThis.matchMedia = function patched(input: string): MediaQueryList {
-				if (input === query) {
-					return {
-						matches: matches as boolean,
-						media: input,
-						onchange: null,
-						addEventListener() {},
-						removeEventListener() {},
-						addListener() {},
-						removeListener() {},
-						dispatchEvent() {
-							return false
-						},
-					} as MediaQueryList
-				}
-				return original(input)
-			}
-		},
-		[TOUCH_PRIMARY_QUERY, is_touch] as const,
-	)
-}
 
 test('page response includes HTTP security headers', async ({ page }) => {
 	const response = await page.goto('/')
